Improved Specific Price with Tax Changes
There is a bug in prestashop that doesn't calculate correctly Specific Prices when taxes are changed in Front Office.

For example, with a 21% taxes when the product costs 100€ and it has an specific price of 10€ tax included total price should be 90€ tax included.
Without taxes should be:
original price: 82,64€
discount: 8,26 €
total price: 74,34 €

The problem is that the discount is always 10€. So i modiffied the way it gets the taxes of product.
With this, it checks the default country and gets its tax rule. When checks the specific_price amount it calculates the current tax_rule and calculates the difference of % into its amount to display and get the correct amount in product price discount.